Move tetris to src root
